requirements.txt
celery==5.2.6
Django==4.0.4
django-debug-toolbar==3.2.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
ipython==8.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyTelegramBotAPI==4.4.0
python-decouple==3.6
simplejson==3.17.6
gunicorn
django-cors-headers
django-otp==1.1.3
qrcode==7.3.1
python-binance==1.0.16
django-import-export==2.8.0
